[processing] move script provider stuff into provider directory
